Improve report format
There's now a possibility to put a `name` and `description` in the
specfile for each dataset and each model.  These will get used to make
the report a little more explanatory.

Also, the legend gets plotted separately, to avoid over-labelling.

And we trim sample names to remove common prefixes, again to improve
the legibility of the labelling