Perf: Move to standard bioconductor image, static renv
Distribute static renv.lock file so that we're more resilient to
overly restrictive version restrictions sneaking in.

Take advantage of R-4.3.2, and a bump in quarto/deno version that
allows us a nicer parse interface, and allows thumbnails.

Rethink the sidebar contents, so that singleton alignments and/or spec
files don't produce unnecessary singleton nesting.