Reorganise 'make' system to flow more naturally
`shared.mk` replaces `global.mk`, and secrets becomes `secret.mk`.
The _module_ version of these will get generated automatically, but
with better dependency-tracking to their originals, and it's only
necessary for a makefile to to reference `shared.mk`, which will bring
with it the secrets. `local.mk` now becomes `module.mk`.