Make 'build' process obsolescent
Rephrase the build logic so that prerequisites outside the current
module are, when they are still present in the worktree, maintained so
that e.g. spec-file changes in docs still propogate dependencies. When
they are no longer there (e.g. if the module has been isolated) then
conditional makefile will ignore them and revert to the within-module
copies that were generated from an initial 'build'