Handle renaming any of the module directory names
Sometimes it's handy to be able to copy `babs/differential` to
e.g. `babs/differential-2` and that breaks the hard-wired way we'd
looked for optionally prerequisite modules.  So now set
e.g. `diff_dir=.` in that folder's makefile, and have `diff_dir` in
the shared.mk only use `wildcard ../differential` if the variable has
not already been set.