Add demo of how to run an isolated differential analysis
`make test/isolated-analysis` will populate that directory with the
minimal set of files that are required to run a differential analysis
that would usually be achieved after running the nfcore pipeline.
Instead, this removes the need for the docs/ingress/nfcore sibling
directories, and is entirely self-contained, but now takes a dds
counts object stored in `extdata/*.rda`, alongside a spec file and a
very minimal `extdata/*.config`.